refactor: split components into features/ and ui/ tiers, make InputCurrency a pure display adapter (#29)
Move account, budget-settings, and budget-user-manager into
src/lib/components/features/, relocate logo.svelte to ui/logo/ and
side-menu.svelte to features/side-menu/, each with an index.ts barrel.
Extend the budget-user-manager barrel to all three components and
switch consumers to explicit barrel imports.

Remove the Money domain coupling from InputCurrency: value is now
cents (number | null) only, converted with plain arithmetic; the
field.as spread support (string values, consumed defaultValue/type
props) is dropped and account-create binds cents explicitly like the
other call sites.

Document the features/ vs ui/ invariant in docs/components.md and add
a Feature module glossary entry to CONTEXT.md.
