+++ CONTEXT.md
@@ -23,3 +23,7 @@ _Avoid_: custom outline, per-component ring values
 **Adapter**:
 A mechanical entry point (remote function or server load) that connects SvelteKit to user-context: it guards auth, validates input, translates form semantics, redirects, and refreshes caches — but produces no values and holds no business rules (see ADR-0002).
 _Avoid_: endpoint, controller, service layer
+
+**Feature module**:
+A component directory coupled to a specific application capability — imports remote functions, orchestrates forms, owns interaction state. Lives under `src/lib/components/features/`, one directory per capability with an `index.ts` barrel; consumers import from the explicit path (e.g. `$lib/components/features/account`). Primitives in `src/lib/components/ui` stay feature-agnostic: no remote functions, no domain logic, at most type-only domain imports.
+_Avoid_: shared component, domain component, widget
