--- src/lib/components/budget-user-manager/index.ts
@@ -1,1 +1,0 @@
-export { default as BudgetUserManager } from './budget-user-manager.svelte';
