+++ src/lib/components/features/side-menu/index.ts
@@ -1,0 +1,1 @@
+export { default as SideMenu } from './side-menu.svelte';
