+++ src/lib/components/ui/input-currency/input-currency.svelte.test.ts
@@ -92,19 +92,6 @@ describe('InputCurrency cent binding', () => {
 		expect(input.value).toContain('0');
 	});
 
-	it('renders string "0" (as from field.as("number", 0)) as formatted value', () => {
-		render(InputCurrency, {
-			props: {
-				currency: 'EUR',
-				intlConfig: { locale: 'de-DE' },
-				value: '0'
-			}
-		});
-
-		const input = screen.getByRole<HTMLInputElement>('textbox');
-		expect(input.value).toContain('0,00');
-	});
-
 	it('submits cent integer value in form data', async () => {
 		const user = userEvent.setup();
 		const form = document.createElement('form');
