+++ src/lib/components/ui/logo/index.ts
@@ -1,0 +1,1 @@
+export { default as Logo } from './logo.svelte';
